Enable Thread Runner in AT
