+++ acceptance-tests/src/test/java/tech/pegasys/eth2signer/dsl/signer/SignerConfiguration.java
@@ -33,7 +33,7 @@ public class SignerConfiguration {
     return hostname;
   }
 
-  public int httpRpcPort() {
+  public int httpPort() {
     return httpRpcPort;
   }
